Desktop/chattan/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==1.4.49
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
Jinja2==3.1.2
aiofiles==23.1.0
python-dateutil==2.8.2
reportlab==3.6.12
